rag_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyPDF2==3.0.1
langchain==0.0.346
langchain-community==0.0.13
langchain-google-genai==0.0.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0